Added more WinMM functions, improved missing import handling, and fixed issues with IUnknown COM interface methods for DirectInput objects.

+++ Include/Xidi/WrapperIDirectInputDevice.h
@@ -34,7 +34,7 @@ namespace Xidi
         BOOL polledSinceLastGetDeviceState;
 
         /// Reference count.
-        DWORD refcount;
+        ULONG refcount;
 
         /// Specifies whether or not to use Unicode (this depends on the application configuration).
         BOOL useUnicode;
